calc.c Update: Array based calculations.
We now have the capability of having more than 2 `int`s and 1 operator
in an equation in the `calc` function. We could, for example, do
`50+4*7` and it would return `78`. Since there is no `printDouble()`
function for now it rounds decimals. We also will need to add the
capability of recognizing if the user uses a decimal point, and changing
order of operations with parentheses so that `(5+5)*6` returns `60`, not
`35`.