Quality Update for list.c
This was a quality check on `list.c`. Since I have added things,
`list.c` is not considered complete until someone else (@plankp) checks
and tests it.
Changes:
- The `list_t` functions all start with `list_` now. This is to remove
any sort of conflict to scripts that have `list.c` included. Functions
like `add` and `remove` might be used again.
- I have added a `list_resize(list_t lst, uint32 capt)` function that
sets `lst->capt` to `capt` and reallocates the data. This is used by
`list_shrink()`, `list_add()` and `list_remove()`. It has a failsafe to
not let it resize to less than `lst->size`.
- I have added an `autoShrinkTrigger` variable for `list_t` that marks
at what threshold we need to shrink if `autoShrink == true`.
- `list_remove()` now takes in an index value and will remove the string
at the index instead of just the last string in the list.
- I noticed a bug in the `list_destroy()` function where deleting a list
in the which the size is not 0 will result in an error. I commented out
a `for` loop and I can't see the error anymore. This might be a
permanent or a temporary fix. More testing needed. To see the error,
uncomment the `for` loop in `list_destroy()` and try to delete a list
with a size greater than 0.
- Added a test for lists in `calc.c` by calling `calc -list` in the
console in Q.