Create mouse.c
Implemented a support for mouse...partially :( I'm currently developing it for my os so thought for sharing the code. Take a look at the code and change it according to your needs. I tried it on qemu and when you move cursor it prints. I'm currently sick and will try to improve it in future.